Filter current user from profiles in current channel
Ref: KWP-83

When a new user enters a team and creates a direct message channel with
another user, Mattermosts getProfilesInCurrentChannel function returns
both users while for existing users it it only returns the other user.
This commit will filter out the current user, so only the other user
will be in the profiles. We can then check if there is only one user in
the channel and a web meeting is possible. (this check needs to be
removed once we will support web meetings with multiple users.